Fixed SwiftUI presentation issue
When triggering or presenting a paywall, if the default value for `isPresented` was `true`, the paywall would not present/trigger. In fact, it would only work in iOS 13 but not 14+, due to differences in listener being used.

Added a changelog entry for this.

This fixes: https://linear.app/superwall/issue/SW-826/check-swiftui-presentation-when-ispresenting-is-true-by-default