Stops SwiftUI paywalls randomly dismissing
Explicitly triggered paywalls and presented paywalls would sometimes randomly dismiss. State changes within the presenting view caused a rerendering of the view which temporarily reset the state of the binding that controlled the presentation of the paywall. This was causing the Paywall to dismiss.

I've fixed this by adding boolean checks before and after presenting. This has been tested within Spy where the issues were occurring. This also happened in a customer's app!