Rejigged the assignment/config management
Removed dispatch queues to handle config, triggers, and blocking/non blocking assignments. This had a lot of race conditions and edge cases. Not really the best solution.

Now, everything waits until config has returned. Then, it checks to see if an assignment call is waiting to be called. If it is, it checks to see if it's blocking. A blocking one will load assignments and then fire delayed triggers. A non-blocking one will just fire of assignments and then the triggers.

`identify(with:)` and `Paywall.configure` both now use the same function when setting the app user id. This can create a blocking assignment call when upgrading static config.

A non-blocking assignment call is created when calling identify from an anonymous user (after reset or on new app install).

Triggers are delayed for both when config hasn't returned but also when waiting for a blocking assignment to complete.

More tests added that tests this stuff.