Check for subscriptionGroupIdentifier in iOS 12+ for purchases
A user shouldn't be offered a free trial if they've already bought a product within the subscription group.
On iOS 12+ we get the subscriptionGroupIdentifier of the product and match that against the subscription group ids of the products already purchased. If no such match exists it shows the free trial text.