Bugfix for presentAgain() and cleaner webview creation
`presentAgain()` wasn't working because the request subject was being finished and so not accepting any further messages. Reworked internallyPresent to take in the last request and statepublisher, so that it can just continue sending items to the original state publisher. The bonus of this is we don't have to resend the identity, which was a bit weird.

This also changes the way webview (and some of it's properties) are created so it's easier to test and doesn't involve as much dependency injection.