Fix for <iOS 14 where the spinner wasn't firing publisher
onReceive wasn't being called because of some weird SwiftUI bug on iOS 14. Adding `.onAppear()` before onReceive worked, possible because it delays the publisher. It's a hack but seems that it's the only way to fix Apple's bug. iOS 16 still works as normal.
Also, consolidated the listeners for movedUp and isHidden into just listening for isHidden