Hide instead of recreate presenting window
This should finally fix the bug. When the window was deallocated (set to nil), viewDidDisappear was called on the presented paywall view controller. This was then accessing window again which caused a simultaneous access crash. Now we don't remove the created window, we just hide it and reuse it when a paywall is next presented.