Fix dismissing issue for paywall with no entitlements
If there are no entitlements belonging to a paywall, e.g. all products are consumable, and that paywall is being shown when the polling of web entitlements returns, then it will close the paywall automatically as a “restore”. In reality it’s not a restore at all. Fixes this with a check that paywall entitlements aren’t empty first.